+++ app/src/main/java/com/axt/FINCALC/lengthActivity.java
@@ -69,8 +69,8 @@ public class lengthActivity extends AppCompatActivity {
                 }
                 else
                 {
-                    double inputNum = Double.parseDouble(input.getText().toString()); // converting input into a double
-                    double meters = 0, kilometers = 0, inches = 0, feets = 0, miles = 0, yards = 0, centimeters = 0;
+                    float inputNum = Float.parseFloat(input.getText().toString()); // converting input into a float
+                    float meters = 0, kilometers = 0, inches = 0, feets = 0, miles = 0, yards = 0, centimeters = 0;
 
 
                     //=================Converting any given unit of time in meters===================//
@@ -79,22 +79,22 @@ public class lengthActivity extends AppCompatActivity {
                     else if (spin.getSelectedItem().toString() == "kilometers")
                         meters = inputNum * 1000;
                     else if (spin.getSelectedItem().toString() == "inches")
-                        meters = (double) (inputNum * 0.0254);
+                        meters = (float) (inputNum * 0.0254);
                     else if (spin.getSelectedItem().toString() == "feet")
-                        meters = (double) (inputNum * 0.3048);
+                        meters = (float) (inputNum * 0.3048);
                     else if (spin.getSelectedItem().toString() == "miles")
-                        meters = (double) (inputNum * 1609.344);
+                        meters = (float) (inputNum * 1609.344);
                     else if (spin.getSelectedItem().toString() == "yards")
-                        meters = (double) (inputNum * 0.9144);
+                        meters = (float) (inputNum * 0.9144);
                     else if (spin.getSelectedItem().toString() == "centimeters")
-                        meters = (double) (inputNum * 0.01);
+                        meters = (float) (inputNum * 0.01);
                     //============Converting meters to other units===================================//
                     kilometers = meters / 1000;
-                    inches = (double) (meters / 0.0254);
-                    feets = (double) (meters / 0.3048);
-                    miles = (double) (meters / 1609.344);
-                    yards = (double) (meters / 0.9144);
-                    centimeters = (double) (meters * 100);
+                    inches = (float) (meters / 0.0254);
+                    feets = (float) (meters / 0.3048);
+                    miles = (float) (meters / 1609.344);
+                    yards = (float) (meters / 0.9144);
+                    centimeters = (float) (meters * 100);
                     //=================Output the results=============================================//
                     meter.setText(String.valueOf(meters));
                     kilometer.setText(String.valueOf(kilometers));
